API Banca En Linea

+++ BancaEnLinea.BC/Modelos/CuentaBancaria.cs
@@ -4,7 +4,7 @@ namespace BancaEnLinea.BC.Modelos
     public  class CuentaBancaria
     {
         public int IdCuenta { get; set; }
-        public int NumeroCuentaBancaria { get; set; }
+        public long NumeroCuentaBancaria { get; set; }
         public TipoMoneda Moneda { get; set; }
         public decimal Saldo { get; set; }
         public TipoCuentaBancaria TipoCuentaBancaria { get; set; }
